add more all.js logic
